swarms-master-rustified/tests/agents/test_tool_agent_rustified.rs
### Conversion Viable: Partially

The provided Python code can be partially converted to Rust, but due to the reliance on the `unittest.mock` library and the `transformers` library, which do not have direct Rust equivalents, some parts will need to be adapted or replaced.

### Limitations and Challenges

*   The `unittest.mock` library is used extensively in the Python code for mocking objects, and Rust does not have a direct equivalent. Rust developers often prefer to design their code for testability rather than relying on mock libraries.
*   The `transformers` library is used to import the `AutoModelForCausalLM` and `AutoTokenizer` classes. This library is specific to Python and Natural Language Processing (NLP), so a Rust equivalent will need to be used or implemented.
*   The `swarms` library and the `ToolAgent` class are also specific to the Python codebase, so their Rust equivalents will need to be implemented.

### Rust Conversion

Below is a partial conversion of the provided Python code into Rust. This example does not include a direct conversion of the `unittest.mock` and `transformers` libraries, as they do not have direct Rust equivalents.

```rust
// Partial Rust conversion

// Define the ToolAgent struct
#[derive(Debug)]
pub struct ToolAgent {
    pub name: String,
    pub description: String,
    pub model: Box<dyn std::any::Any>, // Use a trait object for the model
    pub tokenizer: Box<dyn std::any::Any>, // Use a trait object for the tokenizer
    pub json_schema: serde_json::Value, // Use the serde_json crate for JSON handling
    pub debug: bool,
    pub max_array_length: usize,
    pub max_number_tokens: usize,
    pub temperature: f64,
    pub max_string_token_length: usize,
}

impl ToolAgent {
    // Create a new ToolAgent instance
    pub fn new(
        name: String,
        description: String,
        model: Box<dyn std::any::Any>,
        tokenizer: Box<dyn std::any::Any>,
        json_schema: serde_json::Value,
    ) -> Self {
        ToolAgent {
            name,
            description,
            model,
            tokenizer,
            json_schema,
            debug: false,
            max_array_length: 0,
            max_number_tokens: 0,
            temperature: 0.0,
            max_string_token_length: 0,
        }
    }

    // Update the ToolAgent instance with kwargs
    pub fn update(&mut self, kwargs: serde_json::Value) {
        if let Some(debug) = kwargs.get("debug") {
            if let Some(debug_value) = debug.as_bool() {
                self.debug = *debug_value;
            }
        }
        if let Some(max_array_length) = kwargs.get("max_array_length") {
            if let Some(max_array_length_value) = max_array_length.as_u64() {
                self.max_array_length = max_array_length_value as usize;
            }
        }
        if let Some(max_number_tokens) = kwargs.get("max_number_tokens") {
            if let Some(max_number_tokens_value) = max_number_tokens.as_u64() {
                self.max_number_tokens = max_number_tokens_value as usize;
            }
        }
        if let Some(temperature) = kwargs.get("temperature") {
            if let Some(temperature_value) = temperature.as_f64() {
                self.temperature = *temperature_value;
            }
        }
        if let Some(max_string_token_length) = kwargs.get("max_string_token_length") {
            if let Some(max_string_token_length_value) = max_string_token_length.as_u64() {
                self.max_string_token_length = max_string_token_length_value as usize;
            }
        }
    }

    // Run the ToolAgent
    pub fn run(&self, task: String) {
        // Implement the run logic here
        println!("Running the ToolAgent with task: {}", task);
    }
}

#[cfg(test)]
mod tests {
    use super::*;

    // Define a mock model and tokenizer for testing
    #[derive(Debug)]
    struct MockModel;

    #[derive(Debug)]
    struct MockTokenizer;

    #[test]
    fn test_tool_agent_init() {
        let model = Box::new(MockModel);
        let tokenizer = Box::new(MockTokenizer);
        let json_schema = serde_json::json!({
            "type": "object",
            "properties": {
                "name": {"type": "string"},
                "age": {"type": "number"},
                "is_student": {"type": "boolean"},
                "courses": {"type": "array", "items": {"type": "string"}},
            }
        });
        let name = "Test Agent".to_string();
        let description = "This is a test agent".to_string();

        let agent = ToolAgent::new(name, description, model, tokenizer, json_schema);

        assert_eq!(agent.name, "Test Agent");
        assert_eq!(agent.description, "This is a test agent");
    }

    #[test]
    fn test_tool_agent_init_with_kwargs() {
        let model = Box::new(MockModel);
        let tokenizer = Box::new(MockTokenizer);
        let json_schema = serde_json::json!({
            "type": "object",
            "properties": {
                "name": {"type": "string"},
                "age": {"type": "number"},
                "is_student": {"type": "boolean"},
                "courses": {"type": "array", "items": {"type": "string"}},
            }
        });
        let name = "Test Agent".to_string();
        let description = "This is a test agent".to_string();

        let mut agent = ToolAgent::new(name.clone(), description.clone(), model, tokenizer, json_schema);

        let kwargs = serde_json::json!({
            "debug": true,
            "max_array_length": 20,
            "max_number_tokens": 12,
            "temperature": 0.5,
            "max_string_token_length": 20,
        });

        agent.update(kwargs);

        assert_eq!(agent.name, "Test Agent");
        assert_eq!(agent.description, "This is a test agent");
        assert_eq!(agent.debug, true);
        assert_eq!(agent.max_array_length, 20);
        assert_eq!(agent.max_number_tokens, 12);
        assert_eq!(agent.temperature, 0.5);
        assert_eq!(agent.max_string_token_length, 20);
    }

    #[test]
    fn test_tool_agent_run() {
        let model = Box::new(MockModel);
        let tokenizer = Box::new(MockTokenizer);
        let json_schema = serde_json::json!({
            "type": "object",
            "properties": {
                "name": {"type": "string"},
                "age": {"type": "number"},
                "is_student": {"type": "boolean"},
                "courses": {"type": "array", "items": {"type": "string"}},
            }
        });
        let name = "Test Agent".to_string();
        let description = "This is a test agent".to_string();

        let agent = ToolAgent::new(name, description, model, tokenizer, json_schema);

        let task = "Generate a person's information based on the following schema:".to_string();

        agent.run(task);
    }
}
```

Note: This is a simplified example and does not cover the full complexity of the provided Python code. It serves as a starting point for converting the code to Rust and highlights the challenges and limitations of the conversion process.

The conversion requires replacing the `unittest.mock` library with a Rust equivalent, such as using trait objects to represent the model and tokenizer. The `transformers` library will need to be replaced with a Rust equivalent for Natural Language Processing (NLP) tasks. Additionally, the `swarms` library and `ToolAgent` class will need to be implemented in Rust.

### Dependency Management

To use the `serde_json` crate in your Rust project, add the following dependency to your `Cargo.toml` file:

```toml
[dependencies]
serde_json = "1.0.85"
```